==> code/SangGyeom/03_1_query_test.ipynb <==
import os
import re
import time
from typing import List, Dict, Any, Tuple

# 최신 LangChain API 사용
from langchain.schema import Document
from langchain_community.vectorstores import Chroma
from langchain.embeddings import HuggingFaceEmbeddings
from sentence_transformers import SentenceTransformer
import importlib.util

# 파일 경로를 지정하여 모듈 로드 (파일명이 숫자로 시작하더라도 importlib를 이용하면 임포트가 가능합니다)
spec = importlib.util.spec_from_file_location("query_test", "./03_1_query_test.py")
query_test = importlib.util.module_from_spec(spec)
spec.loader.exec_module(query_test)

# 임포트된 함수 확인
print("임포트된 함수들:")
print(" - load_vector_db:", query_test.load_vector_db)
print(" - advanced_query:", query_test.advanced_query)
print(" - display_grouped_results:", query_test.display_grouped_results)
print(" - run_query:", query_test.run_query)



# 벡터 DB 디렉토리 경로 (환경에 맞게 수정하세요)
db_dir = './db/vector_db'

# 벡터 DB 로드 테스트
vectordb = query_test.load_vector_db(db_dir)
if vectordb is None:
    print("load_vector_db 테스트: 벡터 DB 로드 실패")
else:
    print("load_vector_db 테스트: 벡터 DB 로드 성공")


# advanced_query 함수 테스트
if vectordb is not None:
    sample_query = "테스트 쿼리"
    grouped_results = query_test.advanced_query(vectordb, sample_query)
    print("\nadvanced_query 테스트 완료. 그룹화된 결과:")
    print(grouped_results)
else:
    print("벡터 DB가 로드되지 않아 advanced_query 테스트를 수행할 수 없습니다.")


test_query = "3층 슬라브 작업시 이동중 미끄러짐"
print("run_query 함수 테스트 시작:")
query_test.run_query(db_dir, test_query)


test_query = "배수로 드레인 이동통로 확보상태 미흡, 작업자의 불안전한 상태에서의 이동 작업"
print("run_query 함수 테스트 시작:")
query_test.run_query(db_dir, test_query)

test_query = "방총작업 중 배관호스에 걸려 넘어진 사고이며, 근로자간 의사소통 미흡으로 인한 사고로 판단됨"
print("run_query 함수 테스트 시작:")
query_test.run_query(db_dir, test_query)

test_query = "반복된 갈퀴작업으로 인한 인대 및 근육 손상 추정"
print("run_query 함수 테스트 시작:")
query_test.run_query(db_dir, test_query)
